+++ src/commands/manage-favorites/components/main.svelte
@@ -1,14 +1,15 @@
-<script>
+<script type="text/typescript">
 	import Directory from './directory.svelte';
 	import Button from './button.svelte';
-	import { messageBus } from '../utility/message-bus.ts';
-	import { fromSettings, toSettings, areFavoritesValid } from '../utility/favorites-transform.ts';
-	import { showMessageDialog } from '../utility/dialog-utility.ts';
-	import { asyncData } from '../utility/unicode-data.ts'; // Preload
-	import { vscode } from '../utility/vscode-api.ts';
+	import { messageBus } from '../utility/message-bus';
+	import { fromSettings, toSettings, areFavoritesValid, FavoritesViewData } from '../utility/favorites-transform';
+	import { showMessageDialog } from '../utility/dialog-utility';
+	import { vscode } from '../utility/vscode-api';
 	import { onMount } from 'svelte';
 
-	let favorites = null;
+	import('../utility/unicode-data'); // Start async data loading
+
+	let favorites: FavoritesViewData | null = null;
 	let isDirty = false;
 
 	onMount(async () =>
@@ -19,8 +20,13 @@
 	async function getFavorites()
 	{
 		const favoritesMessage = await messageBus.call('favorites', { type: 'get-favorites' });
-		favorites = fromSettings(favoritesMessage.favorites);
-		favorites.isExpanded = true;
+		favorites = {
+			global: fromSettings(favoritesMessage.global ?? {}),
+			workspace: fromSettings(favoritesMessage.workspace ?? {}),
+		};
+		favorites.global.isExpanded = true;
+		favorites.workspace.isExpanded = true;
+
 		isDirty = false;
 		saveState();
 	}
@@ -28,13 +34,19 @@
 	/** Save settings button. */
 	async function save()
 	{
-		if (areFavoritesValid(favorites) == false)
+		if (favorites == null ||
+			areFavoritesValid(favorites.global) == false ||
+			areFavoritesValid(favorites.workspace) == false)
 		{
 			await showMessageDialog('Multiple directories at the same level have the same name. Cannot save favorites.');
 			return;
 		}
 
-		messageBus.send({ type: 'save', favorites: toSettings(favorites) });
+		messageBus.send({
+			type: 'save',
+			global: toSettings(favorites.global),
+			workspace: toSettings(favorites.workspace),
+		});
 		isDirty = false;
 	}
 	/** Revert settings button. */
@@ -55,11 +67,11 @@
 		vscode.setState({
 			isDirty,
 			favorites,
-		});
+		} as State);
 	}
 	async function loadState()
 	{
-		const state = vscode.getState();
+		const state = vscode.getState<State | null>();
 		if (state != null)
 		{
 			isDirty = state.isDirty;
@@ -71,7 +83,7 @@
 		}
 	}
 
-	function onKeyDown(e)
+	function onKeyDown(e: KeyboardEvent)
 	{
 		if (e.ctrlKey === true &&
 			e.altKey === false &&
@@ -82,6 +94,12 @@
 			save();
 		}
 	}
+
+	interface State
+	{
+		isDirty: boolean,
+		favorites: FavoritesViewData,
+	}
 </script>
 
 <style>
@@ -91,6 +109,9 @@
 		border-bottom: 1px solid;
 	}
 
+	h2 {
+		font-size: 14px;
+	}
 	.fav-tree {
 		display: inline-grid;
 		grid-template-columns: auto auto;
@@ -110,8 +131,15 @@
 <h1>Unicode Favorites</h1>
 
 {#if favorites}
+	<h2>User Settings</h2>
 	<div class="fav-tree">
-		<Directory node={favorites}
+		<Directory node={favorites.global}
+			on:change={onChange}/>
+	</div>
+
+	<h2>Workspace Settings</h2>
+	<div class="fav-tree">
+		<Directory node={favorites.workspace}
 			on:change={onChange}/>
 	</div>